+++ packages/components/src/components/accordion/accordion-styles.spec.ts
@@ -1,22 +1,26 @@
+import { describe, it } from 'vitest';
 import { validateCssAndMatchSnapshot } from '../../../tests/unit/helpers';
 import { getComponentCss } from './accordion-styles';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    ['small', false, true, false],
-    ['small', false, false, false],
-    ['small', true, true, false],
-    ['small', true, false, false],
-    ['medium', false, true, false],
-    ['medium', false, false, false],
-    ['medium', true, true, false],
-    ['medium', true, false, false],
-    ['medium', true, false, true],
-    ['small', false, true, false],
-    ['medium', true, false, false],
-    ['medium', true, false, true],
-    [{ base: 'small', xs: 'medium', s: 'small', m: 'medium', l: 'small', xl: 'medium' }, false, true, false],
-  ])('should return correct css for size: %j, compact: %s, open: %s and sticky: %s', (...args) => {
-    validateCssAndMatchSnapshot(getComponentCss(...args));
-  });
+    ['start', 'frosted', false, false, false, false, false, 'small'],
+    ['start', 'frosted', true, false, false, false, false, 'medium'],
+    ['start', 'frosted', true, true, false, false, false, 'small'],
+    ['start', 'frosted', true, true, true, false, false, 'small'],
+    ['end', 'frosted', true, true, true, false, false, 'small'],
+    ['start', 'frosted', true, true, true, true, false, 'small'],
+    ['start', 'frosted', true, true, true, true, true, 'small'],
+    ['start', 'frosted', false, false, false, false, false, 'small'],
+    ['start', 'canvas', false, false, false, false, false, 'small'],
+    ['start', 'canvas', false, false, true, false, false, 'small'],
+    ['start', 'canvas', false, true, true, false, false, 'small'],
+    ['start', 'surface', false, true, true, false, false, 'small'],
+    ['start', 'none', false, false, false, false, false, 'small'],
+  ])(
+    'should return correct css for alignMarker: %s, background: %s, isCompact: %s, isOpen: %s, isSticky: %s, hasSummaryBefore: %s, hasSummaryAfter: %s, size: %s',
+    (...args) => {
+      validateCssAndMatchSnapshot(getComponentCss(...args));
+    }
+  );
 });