+++ packages/components/src/components/accordion/accordion.tsx
@@ -1,33 +1,44 @@
-import { Component, Element, Event, type EventEmitter, Host, h, type JSX, Prop } from '@stencil/core';
+import { Component, Element, Event, type EventEmitter, forceUpdate, h, type JSX, Prop } from '@stencil/core';
 import type { BreakpointCustomizable, PropTypes } from '../../types';
 import {
   AllowedTypes,
   attachComponentCss,
-  getPrefixedTagNames,
   HEADING_TAGS,
+  hasNamedSlot,
   hasPropValueChanged,
+  observeChildren,
+  unobserveChildren,
   validateProps,
 } from '../../utils';
 import { getComponentCss } from './accordion-styles';
 import {
+  ACCORDION_ALIGN_MARKERS,
   ACCORDION_SIZES,
+  ACCORDIONS_BACKGROUNDS,
+  type AccordionAlignMarker,
+  type AccordionBackground,
   type AccordionHeadingTag,
   type AccordionSize,
   type AccordionUpdateEventDetail,
 } from './accordion-utils';
 
 const propTypes: PropTypes<typeof Accordion> = {
+  open: AllowedTypes.boolean,
+  alignMarker: AllowedTypes.oneOf<AccordionAlignMarker>(ACCORDION_ALIGN_MARKERS),
+  background: AllowedTypes.oneOf<AccordionBackground>(ACCORDIONS_BACKGROUNDS),
+  compact: AllowedTypes.boolean,
+  sticky: AllowedTypes.boolean,
   size: AllowedTypes.breakpoint<AccordionSize>(ACCORDION_SIZES),
   heading: AllowedTypes.string,
   headingTag: AllowedTypes.oneOf<AccordionHeadingTag>(HEADING_TAGS),
-  open: AllowedTypes.boolean,
-  compact: AllowedTypes.boolean,
-  sticky: AllowedTypes.boolean,
 };
 
 /**
- * @slot {"name": "heading", "description": "Defines the heading used in the accordion. Can be used alternatively to the heading prop. Please **refrain** from using any other than text content as slotted markup." }
- * @slot {"name": "", "description": "Default slot for the main content" }
+ * @slot {"name": "summary", "description": "Content for the accordion's summary section. Clicking toggles the accordion open and closed." }
+ * @slot {"name": "summary-before", "description": "Content or interactive elements placed before the accordion's summary section." }
+ * @slot {"name": "summary-after", "description": "Content or interactive elements placed after the accordion's summary section." }
+ * @slot {"name": "heading", "description": "Content for the accordion's heading section. Clicking toggles the accordion open and closed.", "isDeprecated": true }
+ * @slot {"name": "", "description": "Main content displayed when the accordion is expanded." }
  *
  * @controlled {"props": ["open"], "event": "update"}
  */
@@ -38,28 +49,61 @@ const propTypes: PropTypes<typeof Accordion> = {
 export class Accordion {
   @Element() public host!: HTMLElement;
 
-  /** The text size. */
-  @Prop() public size?: BreakpointCustomizable<AccordionSize> = 'small';
-
-  /** Defines the heading used in accordion. */
-  @Prop() public heading?: string;
-
-  /** Sets a heading tag, so it fits correctly within the outline of the page. */
-  @Prop() public headingTag?: AccordionHeadingTag = 'h2';
-
-  /** Defines if accordion is open. */
+  /** Controls whether the accordion is open or closed. */
   @Prop() public open?: boolean;
 
-  /** Displays the Accordion as compact version with thinner border and smaller paddings. */
+  /** Aligns the marker within the summary section. */
+  @Prop() public alignMarker?: AccordionAlignMarker = 'end';
+
+  /** Defines the background color. Use `frosted` only on images, videos or gradients. */
+  @Prop() public background?: AccordionBackground = 'none';
+
+  /** Displays the accordion in compact mode. */
   @Prop() public compact?: boolean;
 
   /**
-   * @experimental Sticks the Accordion heading at the top, fixed while scrolling
+   * @deprecated, will be removed in the next major release. Use the `summary` slot instead.
+   * Controls the heading size in the summary section (only applies when using the `heading` prop or `heading` slot). */
+  @Prop() public size?: BreakpointCustomizable<AccordionSize> = 'small';
+
+  /**
+   * @deprecated, will be removed in the next major release. Use the `summary` slot instead.
+   * Sets the heading text within the summary section. */
+  @Prop() public heading?: string;
+
+  /**
+   * @deprecated, will be removed in the next major release. Use the `summary` slot instead.
+   * Sets the heading tag for proper semantic structure within the page. */
+  @Prop() public headingTag?: AccordionHeadingTag = 'h2';
+
+  /**
+   * @experimental Makes the summary section sticky at the top while scrolling. Only works with `background="canvas"` or `background="surface"`. Not compatible with `summary-before` or `summary-after` slots.
    */
   @Prop() public sticky?: boolean;
 
   /** Emitted when accordion state is changed. */
   @Event({ bubbles: false }) public update: EventEmitter<AccordionUpdateEventDetail>;
+
+  private hasSummary: boolean;
+  private hasSummaryBefore: boolean;
+  private hasSummaryAfter: boolean;
+
+  public connectedCallback(): void {
+    // Observe dynamic slot changes. As soon as CSS selector :has-slotted has better browser support, we can remove this
+    // and use that selector in getComponentCss instead.
+    observeChildren(
+      this.host,
+      () => {
+        forceUpdate(this.host);
+      },
+      undefined,
+      { subtree: false, childList: true, attributes: false }
+    );
+  }
+
+  public disconnectedCallback(): void {
+    unobserveChildren(this.host);
+  }
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
     return hasPropValueChanged(newVal, oldVal);
@@ -67,45 +111,44 @@ export class Accordion {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    attachComponentCss(this.host, getComponentCss, this.size, this.compact, this.open, this.sticky);
 
-    const buttonId = 'accordion-control';
-    const contentId = 'accordion-panel';
+    this.hasSummary = hasNamedSlot(this.host, 'summary');
+    this.hasSummaryBefore = hasNamedSlot(this.host, 'summary-before');
+    this.hasSummaryAfter = hasNamedSlot(this.host, 'summary-after');
 
-    const PrefixedTagNames = getPrefixedTagNames(this.host);
+    attachComponentCss(
+      this.host,
+      getComponentCss,
+      this.alignMarker,
+      this.background,
+      this.compact,
+      this.open,
+      this.sticky,
+      this.hasSummaryBefore,
+      this.hasSummaryAfter,
+      this.size
+    );
+
     const Heading = this.headingTag;
 
     return (
-      <Host>
-        <Heading class="heading">
-          <button
-            id={buttonId}
-            type="button"
-            aria-expanded={this.open ? 'true' : 'false'}
-            aria-controls={contentId}
-            onClick={this.onButtonClick}
-          >
-            {this.heading || <slot name="heading" />}
-            <span class="icon-container">
-              <PrefixedTagNames.pIcon
-                class="icon"
-                name={this.open ? 'minus' : 'plus'}
-                size="xx-small"
-                aria-hidden="true"
-              />
-            </span>
-          </button>
-        </Heading>
-        <div id={contentId} class="collapsible" role="region" aria-labelledby={buttonId}>
-          <div>
-            <slot />
-          </div>
+      <details {...(this.open ? { open: true } : {})}>
+        {/** biome-ignore lint/a11y/noStaticElementInteractions: necessary to enable a controlled state */}
+        <summary onClick={this.onSummaryClick}>
+          {this.hasSummary ? <slot name="summary" /> : <Heading>{this.heading || <slot name="heading" />}</Heading>}
+        </summary>
+        {this.hasSummaryBefore && <slot name="summary-before" />}
+        {this.hasSummaryAfter && <slot name="summary-after" />}
+        <div>
+          <slot />
         </div>
-      </Host>
+      </details>
     );
   }
 
-  private onButtonClick = (): void => {
+  private onSummaryClick = (e: Event): void => {
+    e.preventDefault();
+    e.stopPropagation();
     this.update.emit({ open: !this.open });
   };
 }