+++ packages/components/src/styles/css-variables/font/size/typescale2Xl.ts
@@ -1,0 +1,1 @@
+export const typescale2Xl = 'var(--p-typescale-2xl)' as const;