+++ packages/components/src/styles/css-variables/font/size/typescaleLg.ts
@@ -1,0 +1,1 @@
+export const typescaleLg = 'var(--p-typescale-lg)' as const;