+++ packages/components/src/styles/css-variables/spacing/fluid/spacingFluidXs.ts
@@ -1,0 +1,1 @@
+export const spacingFluidXs = 'var(--p-spacing-fluid-xs)' as const;