+++ packages/components/src/styles/css-variables/spacing/static/spacingStaticSm.ts
@@ -1,0 +1,1 @@
+export const spacingStaticSm = 'var(--p-spacing-static-sm)' as const;